# docs/swagger.json
{
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Document for the Scribe API",
        "title": "Scribe API",
        "contact": {},
        "license": {
            "name": "AGPLv3"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/langs": {
            "get": {
                "tags": [
                    "langs"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/Lang"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/JSONError"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "langs"
                ],
                "parameters": [
                    {
                        "description": "Create body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/lang.CreateBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Lang"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/JSONError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/JSONError"
                        }
                    }
                }
            }
        },
        "/langs/{isoCode}": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "langs"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Lang ISO Code",
                        "name": "isoCode",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/JSONMessage"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/JSONError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/JSONError"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "langs"
                ],
                "parameters": [
                    {
                        "description": "Update body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/lang.UpdateBody"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Lang ISO Code",
                        "name": "isoCode",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Lang"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/JSONError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/JSONError"
                        }
                    }
                }
            }
        },
        "/refresh": {
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "refresh"
                ],
                "summary": "Refresh sources by there types",
                "parameters": [
                    {
                        "type": "array",
                        "items": {
                            "enum": [
                                "rss",
                                "vimeo",
                                "youtube"
                            ],
                            "type": "string"
                        },
                        "description": "Type of sources to refresh",
                        "name": "types",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/Content"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/JSONError"
                        }
                    }
                }
            }
        },
        "/refresh/{sourceID}": {
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "refresh"
                ],
                "summary": "Refresh a given source",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Source ID",
                        "name": "sourceID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/Source"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/JSONError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "Content": {
            "type": "object",
            "properties": {
                "author": {
                    "description": "For feedly article",
                    "type": "string"
                },
                "content": {
                    "type": "string"
                },
                "contentId": {
                    "description": "Youtube or Vimeo ID or Feedly ID",
                    "type": "string"
                },
                "contentUrl": {
                    "description": "Youtube or Vimeo video url or article URL",
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "publishedAt": {
                    "type": "string"
                },
                "rawContent": {
                    "type": "string"
                },
                "rawSummary": {
                    "type": "string"
                },
                "sourceId": {
                    "type": "integer"
                },
                "summary": {
                    "type": "string"
                },
                "thumbnailUrl": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "JSONError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "JSONMessage": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "Lang": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "imageUrl": {
                    "type": "string"
                },
                "isoCode": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "Source": {
            "type": "object",
            "properties": {
                "coverUrl": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "iconUrl": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "lang": {
                    "type": "string"
                },
                "order": {
                    "type": "integer"
                },
                "publishedAt": {
                    "type": "string"
                },
                "refreshedAt": {
                    "type": "string"
                },
                "shortTitle": {
                    "type": "string"
                },
                "skateSource": {
                    "type": "boolean"
                },
                "sourceId": {
                    "description": "Vimeo, Youtube or Feedly ID, depending on the type",
                    "type": "string"
                },
                "sourceType": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "websiteUrl": {
                    "type": "string"
                }
            }
        },
        "lang.CreateBody": {
            "type": "object",
            "required": [
                "imageURL",
                "isoCode"
            ],
            "properties": {
                "imageURL": {
                    "type": "string"
                },
                "isoCode": {
                    "type": "string"
                }
            }
        },
        "lang.UpdateBody": {
            "type": "object",
            "required": [
                "imageURL"
            ],
            "properties": {
                "imageURL": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}